Update day working

+++ new-backend/Services/DayService.cs
@@ -15,12 +15,14 @@ namespace Makro.Services
         private readonly MakroContext _context;
         private readonly ILogger _logger;
         private readonly IMapper _mapper;
+        private readonly MealService _mealService;
 
-        public DayService(MakroContext context, ILogger<DayService> logger, IMapper mapper)
+        public DayService(MakroContext context, ILogger<DayService> logger, IMapper mapper, MealService mealService)
         {
             _logger = logger;
             _context = context;
             _mapper = mapper;
+            _mealService = mealService;
         }
 
         public async Task<ActionResult<IEnumerable<DayDto>>> GetAllDaysByUser(string id)
@@ -55,7 +57,11 @@ namespace Makro.Services
                 m.User = day.User;
                 var mealDto = _mapper.Map<MealDto>(m);
                 var foodDtos = new List<FoodDto>();
-                m.MealFoods.ToList().ForEach(mf => foodDtos.Add(_mapper.Map<FoodDto>(mf.Food)));
+                m.MealFoods.ToList().ForEach(mf => {
+                    var foodDto = _mapper.Map<FoodDto>(mf.Food);
+                    foodDto.Amount = mf.FoodAmount;
+                    foodDtos.Add(foodDto);
+                  });
                 mealDto.Foods = foodDtos;
                 mealDtos.Add(mealDto);
             });
@@ -72,21 +78,49 @@ namespace Makro.Services
             return new ResultDto(true, "Day added succesfully");
         }
 
-        public async Task<ResultDto> UpdateDay(Day day)
+        public async Task<ResultDto> UpdateDay(DayDto dayDto, string userId)
         {
-            _context.Entry(day).State = EntityState.Modified;
-            await _context.SaveChangesAsync();
-            return new ResultDto(true, "Day updated succesfully");
-        }
-
-        public async Task<ResultDto> DeleteDay(int id)
-        {
-            var day = await _context.Days.FindAsync(id);
+            var day = await _context.Days.Where(d => d.UUID == dayDto.UUID)
+                .Include(d => d.User)
+                .Include(d => d.Meals)
+                .FirstOrDefaultAsync();
 
             if (day == null)
             {
-                _logger.LogDebug("Day not found with id: ", id);
+                _logger.LogDebug("Day not found with UUID: ", dayDto.UUID);
                 return new ResultDto(false, "Day not found");
+            }
+
+            if (day.User.UUID != userId)
+            {
+                _logger.LogError("User with UUID ", userId + " tried to modify day which belongs to " + day.User.UUID);
+                return new ResultDto(false, "Unauthorized");
+            }
+
+            day.Meals.ToList().ForEach(m => _context.Meals.Remove(m));
+            day.Meals = _mealService.UpdateMeals(dayDto.AllMeals);
+            day.UpdatedAt = DateTime.Now;
+
+            _context.Entry(day).State = EntityState.Modified;
+            await _context.SaveChangesAsync();
+
+            return new ResultDto(true, "Day updated succesfully");
+        }
+
+        public async Task<ResultDto> DeleteDay(string id, string userId)
+        {
+            var day = await _context.Days.Where(d => d.UUID == id).Include(d => d.User).FirstOrDefaultAsync();
+
+            if (day == null)
+            {
+                _logger.LogDebug("Day not found with UUID: ", id);
+                return new ResultDto(false, "Day not found");
+            }
+
+            if (day.User.UUID != userId)
+            {
+                _logger.LogError("User with UUID ", userId + " tried to delete day which belnogs to " + day.User.UUID);
+                return new ResultDto(false, "Unauthorized");
             }
 
             _context.Days.Remove(day);
